requirements.txt
Flask
Flask-Cors
PyYAML
python-dotenv
requests
google-cloud-speech
